requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
build==0.10.0
    # via pip-tools
click==8.1.6
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
django==4.2.3
    # via
    #   -r requirements.in
    #   django-phonenumber-field
django-phonenumber-field[phonenumbers]==7.1.0
    # via -r requirements.in
packaging==23.1
    # via build
phonenumbers==8.13.17
    # via django-phonenumber-field
pip-tools==7.1.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-decouple==3.8
    # via -r requirements.in
sqlparse==0.4.4
    # via django
tzdata==2023.3
    # via django
wheel==0.41.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
